fix: reorder image processing order, add debug outputs for all steps fix: improve duration parsing regex feat: allow float fps values fix: properly handle empty midi frames feat: add interpolation_ongoing variable to context fix: remove seemingly unnecessary tensor op when warping

+++ t2v/config/root.py
@@ -11,7 +11,7 @@ class RootConfig:
     """
     The root of the config file
     """
-    frames_per_second: int
+    frames_per_second: float
     scenes: List[Scene]
     mechanisms: List[MechanismDefinition]
     width: int
